Add timeouts to boot and shut down
Assume that we are going to use a read-only filesystem to avoid data
corruption which is not a guarantee that the SD Card/eMMC won't corrupt
but almost.

This is just in case of a catastrophic failure on boot and also on shut
down to avoid running of out battery.